move preparation logic to MvRewritePreprocessor
